Crate rtf_docgen

Functions

find_fp_variants 🔒

Returns all the IDs for the implementations within the variants of the FileProvider enum. These are the concrete top-level items that we want to document.

get_docs 🔒

get_name 🔒

lookup_item 🔒

main 🔒

write_struct_docs 🔒

Writes the documentation for each variant and its fields. Does not include documentation for fields of fields at this time.

write_table_of_contents 🔒

Writes the table of contents for all the File Providers that will be documented on this page.