Crate rtf_docgen

Statics

FILE_PROVIDER_IDS 🔒

Functions

document_inner_fields 🔒

document_inner_variants 🔒

document_item_content 🔒

Writes out the documentation for a struct’s fields or an enum’s variants

find_fp_variants 🔒

Returns all the IDs for the implementations within the variants of the FileProvider enum. These are the concrete top-level items that we want to document.

generate_toc_link 🔒

Generates the table of contents link for a File Provider

get_docs 🔒

get_name 🔒

lookup_item 🔒

main 🔒

write_file_provider_docs 🔒

Writes the top-level documentation for each file provider and recursively documents their fields.

write_table_of_contents 🔒

Writes the table of contents for all the File Providers that will be documented on this page.